Pre-pull go-through
Cleaned up some things, ready to merge with master.
