+++ StevenUniverseUnity/Assets/Scripts/Characters/Customization/Skill.cs
@@ -1,13 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
+﻿
 namespace StevenUniverse.FanGame.Characters.Customization
 {
     [System.Serializable]
     public class Skill
     {
+
+        // Placeholder for a skill framework.
+
         public Skill()
         {
