--- StevenUniverseUnity/Assets/Scripts/Entities/Customization/UnitStats.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace StevenUniverse.FanGame.Characters.Customization
-{
-    [System.Serializable]
-    public class UnitStats
-    {
-        public UnitStats()
-        {
-
-        }
-    }
-}
